+++ DependencyInjection/c975LXliffExtension.php
@@ -28,12 +28,5 @@ class c975LXliffExtension extends Extension
             new FileLocator(__DIR__.'/../Resources/config')
         );
         $loader->load('services.yml');
-
-        $configuration = new Configuration();
-        $processedConfig = $this->processConfiguration($configuration, $configs);
-
-        $container->setParameter('c975_l_xliff.source', $processedConfig['source']);
-        $container->setParameter('c975_l_xliff.languages', $processedConfig['languages']);
-        $container->setParameter('c975_l_xliff.rootTranslations', $processedConfig['rootTranslations']);
     }
 }